Add the CAN_FD support
Update of the SendFD function to have Send() for standard CAN and SendFD() for CAN_FD frame.
Update of the buffer size to be compatible with 64 bytes data (CAN_FD)
Auto detection of the CAN_FD request if the number of data is > 8 bytes.
Documentation (send.html) have been updated also

!!! Need to be updated with SOCKETCAN update also. In parallel, the SOCKETCAN not include the support of the CAN_FD and updated have be send to the owner of the SOCKETCAN to add the support of the CAN_FD.
This code is retro-compatible even if SOCKETCAN is not yet updated. 
The old flow don't need to be updated and the code have been develop in this way to have no effect to the current version of this contrib.